MMS creates a new thread when receiver is not in the Message thread
Bug 7334111

The bug is due to the phone not knowing its own number and hence, not being
able to remove the phone's own number from the list of recipients. The simple
fix for one-to-one chats is to not add the TO recipients when there's only
one recipient, because we know that's going to be ourself.

Change-Id: Iec54949e53c350974b8291e389f61b323f1b23f2