Force mobile data enabled on lollipop
This actually will not work correctly. Previously, there was a simple method
which could be grabbed through reflection and invoked to enable mobile data
if it is disabled on a device... This method no longer exists on lollipop.

I have invoked a different method through reflection that should be able to
handle the task of enabling data now, but it is locked to third party app
use because of its requirement of a permission only available to system apps.