Buildable for Android Marshmallow
Currently, the sending will fail. There seems to be changes in Marshmallow that
do not grant third party apps the permission CHANGE_NETWORK_STATE.
https://code.google.com/p/android-developer-preview/issues/detail?id=2993.

Hopefully this will be corrected before the final release as this library
needs this functionality for ConnectivityManager.RequestNetwork() to work
correctly and send over mobile data instead of wifi.

There is still a lot of work to be done on the Marshmallow front. Since the
Apache HTTP library is now deprecated, I need to update all sending methods
to use URLConnection instead. Android 5.1+ uses this functionality, so it
should be a simple task of bringing it in and backporting it to use on all API
levels.